pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "career-co-pilot-pro"
version = "0.1.0"
description = "AI-powered job application automation system"
requires-python = ">=3.10"
dependencies = [
    # 0.135+ required for Starlette 0.46+ (Router no longer accepts on_startup/on_shutdown)
    "fastapi>=0.135",
    "uvicorn[standard]>=0.22",
    "celery[redis]>=5.3",
    "pydantic>=1.10",
    "redis>=4.5",
    "langchain-openai",
    "langgraph",
    "streamlit",
    "pandas",
    "reportlab",
    "bs4",
    "requests",
    "python-dotenv",
    "PyMuPDF",
    "apify-client",
    "scikit-learn",
    "joblib",
    "nltk"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy"
]
# LinkedIn MCP + Playwright apply path (same as `mcp`; alias for docs / review clarity)
apply = [
    "mcp",
    "fastmcp",
    "playwright",
]
mcp = [
    "mcp",
    "fastmcp",
    "playwright"
]
full = [
    "mcp",
    "fastmcp",
    "playwright"
]
# Typical production install: tracker DB, JWT, migrations, optional S3 + metrics
production = [
    "psycopg2-binary>=2.9",
    "PyJWT>=2.8",
    "alembic>=1.13",
    "sqlalchemy>=2.0",
    "boto3>=1.34",
    "prometheus-client>=0.20",
]
auth = [
    "PyJWT>=2.8",
]
postgres = [
    "psycopg2-binary>=2.9",
]
migrations = [
    "alembic>=1.13",
    "sqlalchemy>=2.0",
]
s3 = [
    "boto3>=1.34",
]
metrics = [
    "prometheus-client>=0.20",
]

[tool.setuptools.packages.find]
include = ["app*", "agents*", "dashboard*", "config*", "services*", "providers*", "ui*", "mcp_servers*"]
